feat(home): embed a playable Flavor Town round on the homepage
The round moves out of GuessLive.Play into a shared GameComponent.
/flavor-town keeps its behavior (random card per round, sticky mobile
guess bar, in-place "Play again"); the homepage embeds the component in
:embedded mode — the deterministic daily teaser card as everyone's
shared puzzle, an inline guess bar, and a "Play more" link into
/flavor-town once the round ends. The static quote teaser it replaces
is removed from HomeLive.
